Fix dir/vc lines of PS1
Fix color codes to work correctly when not used statically. Now that PS1
format changes based on context, the dir/VC info is clogged with escaped
color codes.

To fix:
  - Add `ansi_color_*` variables to `_colors_setup` for color codes
    without PS1 escape brackets.
  - Change some `echo` commands to `echo -e` to allow them to process
    escape sequences.
  - Change to use the `ansi_color_*` vars in the dir/VC info section of
    PS1.

Dir/vc are now evaluated every prompt so that bash's `\w` is not in
charge of directory path formatting. This is partially in prep for a
future 2 color directory scheme while in version control repos.